+++ src/qorus_creator/InterfaceInfo.ts
@@ -4,7 +4,7 @@ import * as flattenDeep from 'lodash/flattenDeep';
 import { qorus_webview } from '../QorusWebview';
 import { default_version, QorusProjectCodeInfo } from '../QorusProjectCodeInfo';
 import { defaultValue, configItemFields } from './config_item_constants';
-import { hasConfigItems, deepCopy } from '../qorus_utils';
+import { hasConfigItems, deepCopy, capitalize } from '../qorus_utils';
 import { t } from 'ttag';
 import * as msg from '../qorus_message';
 
@@ -12,8 +12,10 @@ import * as msg from '../qorus_message';
 export class InterfaceInfo {
     private code_info: QorusProjectCodeInfo;
     private iface_by_id = {};
-    private last_conf_group: string;
     private are_orig_config_items_set: boolean = false;
+
+    private last_conf_group: string | undefined;
+    private last_other_if_kind: string | undefined; // one of ['Group', 'Event', 'Queue']
 
     constructor(project_code_info: QorusProjectCodeInfo) {
         this.code_info = project_code_info;
@@ -76,6 +78,14 @@ export class InterfaceInfo {
         return this.last_conf_group;
     }
 
+    get last_other_iface_kind(): string | undefined {
+        return this.last_other_if_kind;
+    }
+
+    set last_other_iface_kind(other_iface_kind: string | undefined) {
+        this.last_other_if_kind = other_iface_kind && capitalize(other_iface_kind);
+    }
+
     updateConfigItemValue = ({iface_id, iface_kind, name, value, level, parent_class, remove, is_templated_string}) => {
         this.maybeInitIfaceId(iface_id, iface_kind);
         if (!level) {
@@ -94,7 +104,7 @@ export class InterfaceInfo {
             level = 'local';
         }
 
-        const parsed_value = jsyaml.safeLoad(value);
+        const parseIfComplex = item => ['list', 'hash'].includes(item.type) ? jsyaml.safeLoad(value) : value;
 
         const templated_key = level === 'global'
             ? 'is_global_value_templated_string'
@@ -119,10 +129,9 @@ export class InterfaceInfo {
                 let item;
                 if (index > -1) {
                     item = this.iface_by_id[iface_id]['config-item-values'][index];
-                    item.value = parsed_value;
+                    item.value = parseIfComplex(item);
                 } else {
-                    item = {name, value: parsed_value};
-                    this.iface_by_id[iface_id]['config-item-values'].push(item);
+                    msg.error(t`ConfigItemNotFound ${name}`);
                 }
 
                 item[templated_key] = is_templated_string;
@@ -139,7 +148,7 @@ export class InterfaceInfo {
                         item['remove-global-value'] = true;
                     }
                 } else {
-                    item[level + '-value'] = parsed_value;
+                    item[level + '-value'] = parseIfComplex(item);
                     item[templated_key] = is_templated_string;
                 }
             }
@@ -277,12 +286,19 @@ export class InterfaceInfo {
     getConfigItem = ({iface_id, name}) => {
         const config_items = iface_id && this.iface_by_id[iface_id] && this.iface_by_id[iface_id]['config-items'];
         const config_item = (config_items || []).find(item => item.name === name);
+        if (!config_item) {
+            return;
+        }
 
         let item = { ... config_item };
         item.type = item.type || defaultValue('type');
         if (item.type[0] === '*') {
             item.type = item.type.substr(1);
             item.can_be_undefined = true;
+        }
+
+        if (item.default_value !== undefined && ['list', 'hash'].includes(item.type)) {
+            item.default_value = jsyaml.safeDump(item.default_value).replace(/\r?\n$/, '');
         }
 
         const message = {
@@ -396,6 +412,10 @@ export class InterfaceInfo {
         if (!iface_id) {
             return;
         }
+        if (!['workflow', 'job', 'service', 'class', 'step'].includes(iface_kind)) {
+            return;
+        }
+
         this.maybeInitIfaceId(iface_id, iface_kind);
 
         this.code_info.waitForPending(['yaml']).then(() => {