+++ src/qorus_utils.ts
@@ -94,3 +94,5 @@ export const makeFileUri = (filePath: string) => 'file://' + filePath;
 export const getFilePathFromUri = (uri: string | Uri) => typeof uri === 'string' ? uri.slice(7) : uri.fsPath;
 
 export const deepCopy = (obj: any) => JSON.parse(JSON.stringify(obj));
+
+export const capitalize = str => str[0].toUpperCase() + str.substr(1);