+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.0.4
+
+#### Bug fixes
+
+-   mappers used in class connections are now added automatically to the mappers tag
+-   fixed displaying the workflow keylist tag
+-   editing groups, events and queues is now possible
+-   fixed handling default config item values of complex types
+-   interfaces with class connections can now be re-created
+-   fixed misinterpreting of string values that look like yaml code
+-   it is now possible to set 0 for int and float config items
+-   strings starting with `-` are no longer misinterpeted as lists in config items
+-   the add new field button is no longer replaced by delete button when hovering mapper field
+-   folder tree list is no longer hidden after 4 seconds when choosing a folder
+
 ## 1.0.3
 
 #### Bug fixes